refactoring gui
It is not good to manage the gui of an application with a gen_event
event_handler.
The main problem is that the gui need to manage "out of band" messages
(which results in an handle_info/2 invocation) and these messages are
sent to the gui with the "!" operator (these messages are events from
wxWidget, eg. the window has been closed).
This results in sending not complying OTP messages to the gen_event
process, which forwards them to all the registered event_handlers, and
among these handlers there is also the gui handler.
But this is a dirty approach because it results in forwarding all the
gui related messages (wxWidget stuff) also to the other handlers which
are only interested in domain events.